added: Generate multi-patch default geometries in 2D/3D.
In the input file, use:
<geometry nx="# patch in X" ny="# patches in Y" nz="# patches in Z"/>

This allows for generation of multipath squares/cubes.
The topology for the total model is generated.
If you give the 'sets=true' attribute, default topologysets are generated.

In 1D:
- Vertex1,2 - Vertices
- Boundary = Corners = Collection of the two vertices

In 2D:
- Edge1..4 = Sides of square
- Boundary = Collection of all edges
- Vertex1..4 = Corner of square
- Corners = Collection of all vertices

In 3D:
- Face 1..6 = Faces of cube
- Boundary = Collection of all faces
- Edge1..12 = Edges of cube
- Frame = Collection of all edges
- Vertex1..8 = Vertices of cube
- Corners = Collection of all vertices

Periodicities can also be optionally generated.

In the input file use
<geometry periodic_x="true" periodic_y="true" periodic_z="true">

The latter part is a subject for change.